i3cf: Revert `focus_on_window_activation` setting
Was met with a very specific, very likely to happen use-case that led me
to revert this issue. Simple: start *virtually any* application that
creates a window. They'll most likely immediately request focus.

I like to start most, if not all, applications in the background, if
possible, without them yanking me into the workspace they are assigned
to.