Break mutt and neomutt more apart, prefer mutt by default
This isn't quite the same as vim vs. nvim, since neomutt, from what I
gather, is just a fork of mutt + patches that could possibly be
mainlained.  Nonetheless, these projects won't probably ever merge, but
since patches are going in mutt, neomutt can be viewed as somewhat of
staging grounds for mutt features.

It's a bit of a mess, but I think I'll just stick with this layout.
Regardless, neomutt should work perfectly fine with this as well, seeing
it can also be considered somewhat of a superset of mutt.  Compatibility
shouldn't be an issue.