volume: surrender to pulseaudio, overhaul script
Pretty nifty, if I say so myself.  Don't think it's the fastest thing in
the world...  If I ever get the patience of making this a full-fledged
binary to reap some more performance some day, I just might.  We'll see.

As part of the overhaul, it supports both sinks and sources, meaning it
can control microphone volumes too.