sway: fix scratchpad situation
Use `termite`'s capabilities to set a terminal's title instead of using
the previously used `--name` option.  That would set the `app_id` in
sway's window tree, which seems like something that shouldn't be done.
As such, one can detect these windows as "children" of termite.  Which
now that I think about it, it is leaking termite everywhere… hmm :thinking: