i3b(updates): `true`/`false` are functions, fix pkg count
`true` and `false` usually either depend on the shell being run or are
supplied by the system as binaries.  Naturally, spinning up a process,
no matter how small, fast, and efficient the program being run is, is an
overhead.  This is nit-picking, but can become noticeable in larger
scenarios.  I guess it's a good habit to have.