git: default branch name for new repos to `main`
Yes, I know.  It's pandering at this point, but the main takeaway here
(no pun intended) is the `main` is effectively shorter to type, and
whatever minor benefit I can get from it, I'll reap it.  Now I just need
to get that hardwired as opposed to `master`, which is still pretty
entrenched.