sway: leave screen locking/suspend to my better judgement
Without proper idle inhibition support for wayland, it's more of an
annoyance than a cushion when media is glaringly playing and the screen
locks.  To make matters worse, there's a bug in AMD's drm drivers that,
once the screen shuts down, things just won't work when trying to turn
it back on -- experienced complete process freeze/shutdown, got booted
off my sway session 100% of the times it happened.  Year of the Linux
desktop when?  Kappa