volume: detect default sink/source based on `pactl info`
The default device might not always be `RUNNING`.  E.g. for the sake of
saving power, it might switch to a `SUSPENDED` state if no audio is
being fed to it.  `pactl info` explicitly returns the names of the
default devices, making it a more reliable source of information, if not
considered a source of absolute truth in that regard.